mnist_django/requirements.txt

tensorflow==2.8.3
Pillow==8.2.0
numpy==1.19.2
